New layout in UI
Worked on this for about 1 hour to get a new UI and added some of the appropriate text. However the text does not all fit on to eech panel anympore. So we will have to think of another way to display it .  Also Created a new animation for the cube to zoom in whenever you start the application.