Created new UI menu with Palm Interaction
I have been working on this for the past 2  hours and have got the basics down. When the user opens up their left palm a button will appear on their palm. When they click that the menu will open to the right of their left hand. This will then move with their hand. I am still working on getting it to reset when the hand closes. As well as, figuring out what will be in the menu.