Cards now stay in the world upon hover over them
I have spent the last hour doing research into how to get the cards or flyers that will be displayed in the right palm menu to stay in the world once generated in front of you and not on your hands. I have now been able to do this with the .SetParent function inside of Unity engine. I then chose the object TourStopCardHolder to have its new parent set to mixed reality scene content. This can be seen in the SlideIncube script but will be moved in an update to a general script with the name of cardcontroller