Created all of the info cards for Asphalt Lab
I have spent the last 2 hours creating the information cards that will be able to be selected from the right palm menu. Each of these cards will be a flyer or basically a cube that once you gaze at it with your hand or eyes it will then stick to the real world and not your hand. Once this was done I went through and added all of the data that Dr. Bill Buttlar gave me for his lab. Then I had to organize it on to the cards. I had to make all of the cards bigger as now they are two hands up and across. Compared to being only 1 hand before. Some of these cards feature photos and some have a tab system because there was a lot of text. Since there is so much text I am not sure what we can do to make this feel more fluid. We have been looking into animations and lip syncing but have not found a good solution yet.