Made the buttons have Gold even when built out
The issue we had was that the buttons where only temperorairly gold when using the play mode in Unity. However through creating my own material and seting it to the same setting as the MRTK button material I am able to add this new MizzouGold Material to all of the rectamgular objects. I still need to make a new material that is for circular objects