Made menus stay in the world once opened but still broken
I have spent the last 1.5 hours wokring on getting the menu to stay in the real world. I was successfull with this however where I strugled was getting the menu to come back to the nad when the open button was clicked. I  am in the midle of working on getting the open buttons to detect if the menu is open that way when you look at your hand and the menu is already in the real world the button will be the close button not the opne button. However the buttons are not going away after the hand is deactivated. A simple fix might be in the logic structure I have going on right now and only call the checker function on activate and use the deactive when the hand is gone for both buttons