Menu Functionality Built
I have spent the last 3 hours completing the menu functionality for the UI. Now when a user hits either palm button the menu will open and will be placed in the real world. Instead of where in the past the menus where attached to the users palms now they are stuck exactly where the palms where when the button is clicked. However they are not actualy stuck as I have added a black ball that can be used as a graber to manipulate the menu around. I have also added an X to close out the menu once it is open. On the left  palm I have also added back buttons to go back and forth between the menus. This was not needed for the right as there is only ever one menu page on it. Once this was done I am now working on getting the clipping bounds to work  so that the scrolling items can not be seen off of the back plate. As well as, making it so that when the user clicks open again on their palm that the menu is re-generated where there palm is not where the menu last was.