UI More interactive
I have spent the last hour puting together the portion of the UI where the menus will reposition back to where you open them. For example I open one menu then walk away. I look down at my palm and close it. When I open it again the menu needs to be by my palm. This is the issue I solved here in the menu manager script and on the close buttons as well as scene manager objects is where all of my changes can be seen.