New UI
I have been working on desinging this new UI since 10am this morning. I took breaks at 12-1 then again 2-3 then again 5-7. Leading my total hours worked leaving the total amount of hours worked around 5. Give are take some union breaks here and there. The reasoning for such a large amount of time was because I had to reimagine how the user would be interacting with these cards. While also trying to emulate as much as I could from Mizzou branding. I then had to re-think our 4 cards that we had as the professors from other labs have not got back to us on those emails. The new sections are General (describes the purpose of the lab), news (list any news articles or media that was covered on the lab), Director (this is a short bio of the director(s)), Faculty (this is short bio / list of responsibilities for the faculty). Then I created a who new card and menu system for the tour stops. Instead of populating the world with many cards there will now only be one card showing at a time. With each of the new categories having two cards. A header that gives a short blurb as to what the card is about. The detail card I a text card that shows the user more details about that specific category. With the header card containg a photo at the top with a mizzou branding style being emulated. The menu system instead of being a scrolling list is now a radio button vertical menu system that can be found on the right side of the location card. Once the user hits the open button the who card system will open and then they can togel the radio buttons on and off to see each category. The functionality of this system has not been put into place yet as I wanted it tool good first. What this means is that there will now only be one exit button on the left palm menu to leave the stop. As well as, one close button on the discovery menu to close out of the cards. This makes things more simple and has less going on at once to not over complicate the user.