Tour Stop UI complete
I have spent the last 3 hours completing the tour stop UI. This new UI is based off of Mizzou's web UI for identity standards. I have also cut down the number of topics that there where from 8 to 4. The new topics now are general/purpose, News , Faculty , and directors. These will be applied to all of the Lab location. However for the library and café we are going to need to change up the stops to be more reprehensive of their data. I.E. a library doesn't have a director. With that said at the moment when the user looks at their palm the menu will open up and then the user can select between the 4 tabs on the right hand side of the menu. After this the user can close the menu with their palm. If the user where to leave the menu in one spot and close it and re-open it in a new location the menu will change its position to the new location instead of sticking in the old spot.