UI fin
I have been working on this basically from 9-2am with breaks for food and socializing. I will write up a full report tomorrow. Basically the UI is finished and I am working on getting the cards to autogenerate. Well I mean they do autogenerate it is just that I need the data to populate them. I still haven't figured out what to do for the library and pec cards. Other than that I have the three basic scene built out and the you are able to move between the scene with the menu buttons and by selecting a portal. The mode will update itself to the name of the scene. Pretty much everything is done besides the audio recording and getting the data. THE ONE THING THAT IS LEFT: still need to decide on the best way to make the tutorial but for now bed.