Scenes are updated
I have gone through and changed all of the white text on the yellow buttons to being black text . I then made all of the grabbable objects like the MAP and CARDS to smaller so that you don't move them when you are trying to select a button. I also added a lock button to all of the move able objects.