Update UI
I have spent the last 4 hours going about making changes to the UI. I have made the following changes to the UI. 1. I fixed the issue that was having the tour stop cards not being able to stick to the world. This was a simply fix and I just forgot to drag the menu into the script object on Unity. The next change that I made was resizing and adding handles to every object that is moveable. In the past the grip area was just the card. Now that each object has a handle that will light up gold when the user is hovering or moving the object it helps them understand more that this object is moveable. I then made it so that the handle will disappear when the locking feature is enabled. This makes it better for the user to see that the object is indeed locked. The next thing that I did was create game object for all of the buttons to go in. This is just for organizations and helps the other team integrate vuforia faster. Then I changed the right palm menu card that appears in the tutorial scene. Previously it was just an example of what one of the stops would look like. Now the cards explain how to maneuver and use the tour stop card menu system. I then created two new tour stop card menus for the library and the café. These are going to be hard coded in as they are unique in their categories and will not be auto populated.  Next I need to adjust the auto populating script to turn on and off these separate menus based on if the correct stop is selected. As well as, populate the original menu or the lab menu with the information for the labs. All of the lab menus will have the same categories.