UI Update
I have spent the last 4 hours updating the UI again for the tour. What I did this time was add in new scripts that allow for the common area menus like the cafe to be activated once their corresponding button is clicked. Then once I had this script up and working I went through and created another script called Dynamic stops. This script purpose is to make all of the cards auto generate the text when their button is clicked. This makes it so that the lab menu can just be one menu and the text will be dynamically placed into it. Once this was done I went through a tested the whole application over again to make sure that there where no errors. Which I did not find any errors. This process was tedious and long because I had to go through an rename a bunch of game object so that it would be easier to understand. Not only that I went and made the one script then it broke all the other menus so I had to create the new script LnCtab manager. Where I was trying to just use the old tabmanger it did not like this because the lab and cafe menus are only in discover. While tab manger is being used in the tutorial scene as well. One thing that I could not figure out is why there is this black film over all of the menu cards. I am hoping that after opening and closing Unity this will be gone.