Audio Buttons Added
I spent the last 4 hours working on this project again today. What I did was first go through all of the mizzou websites I could find. With the intentions of finding new header photos for the cards. I found some great images and used them on all the cards keeping a consistency with the number category. Meaning that if the third category on one card was the columns it would be that on every card. This helps the viewer know that they have switched cards and what number category they are on. That is if they can't tell by the title and radio button being filled. The next thing I did was go through all of the materials and change them to be a Mixed reality / standard material. This seemed to fix the dark tint issue that I was facing yesterday. However the colors still seem off and I do not know what to do about it. The next thing that I did was add to the scrip all of the text that the other members have provided me with. Here you will see that it is not all filled out that is because they are still working on it. Then I figured out that I would only need to be changing two photos on the Lab menu dynamically. With these photos being the lab header of the general tab and the director headshot on the director tab. I added in the image objects into the script and had their material update based on what lab was selected. After this I then started making all of the audio buttons. But before I could start I had to go to word and get the audio on and off symbols make them mizzou gold, then remove the background from the images. Import them into unity and turn them into a cut out material. From there I created many prototypes of how the the button object should be structured. Then final settled on two buttons with an icon outside of the button. Then these buttons are found inside of two parents call on and off. So in on there is an on button and icon and same for off but with the off button and icon. When one is clicked the other is turned off. The next step is to get these buttons to play audio based on the lab that is selected. As well as, be able to stop the audio before it is done.