Audio Dynamicaly added
I have spent the last 2 hours working on the audio portion of the UI. The first thing that I did was add in the audio files that Tripp had created. We are using these as a indicator to when you switch scenes. So once the scene is changed a transition audio will play at the start of the scene one time. Then once this was done I went ahead and added in audio sources for the Tutorial. I then record them and added them to the project and assigned them to the sources I created. How the audio works is when the on button is click the button will call on the audio source object and play the audio. Once the off button is clicked the button will pause the audio source. Then when the card is changed the audio will stop and not play until the button is clicked again. I then applied this to the lab, library, and cafe menus. However inside of the dynamic stops script I created two audio sources objects for each of the audio sources on the category. I then created 8 audio clips for each lab. Then inside of the same function that assigns the corresponding information for the lab I assigned those audio clips to the audio sources. Now the audio has been dynamically updated based on the user selection. PLEASE NOTE: I am still waiting on all the clips and text so it is not complete yet. Once it is then we are done!!